+++ Sources/PixelMemory/PixelMemory.swift
@@ -1,5 +1,5 @@
 import PixelCore
 
 public enum PixelMemory {
-    public static let version = "0.0.0"
+    public static let version = "0.1.0"
 }