+++ Sources/PixelRemote/PixelRemote.swift
@@ -1,5 +1,6 @@
 import PixelCore
 
 public enum PixelRemote {
-    public static let version = "0.0.0"
+    public static let version = "0.1.0"
+    public static let protocolVersion = 1
 }